# openapi/pdf-server.openapi.yaml
openapi: 3.0.0
info:
  title: PDF Generator API
  version: 1.0.0
  description: API zur Generierung von PDF-Dokumenten aus HTML-Templates mit Handlebars, Nunjucks oder Liquid.
servers:
  - url: http://localhost:3000/api
    description: Lokaler Entwicklungsserver

paths:
  /generate/donc:
    post:
      summary: Generiert eine DONC-PDF mit Handlebars-Template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                name: Max Mustermann
                date: "2025-05-09"
      responses:
        200:
          description: Erfolgreiche PDF-Rückgabe
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        500:
          description: Fehler bei der PDF-Erstellung

  /generate/{templateName}:
    post:
      summary: Generiert ein PDF aus einem dynamischen Template (Handlebars, Nunjucks, Liquid)
      parameters:
        - name: templateName
          in: path
          required: true
          schema:
            type: string
          description: Der Name des Template-Verzeichnisses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        200:
          description: Erfolgreiche PDF-Rückgabe
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        400:
          description: Ungültige Template-Engine
        404:
          description: Template nicht gefunden
        500:
          description: Fehler bei der PDF-Erstellung

  /templates:
    get:
      summary: Gibt eine Liste der verfügbaren Templates zurück
      responses:
        200:
          description: Erfolgreiche Rückgabe der Template-Namen
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        500:
          description: Fehler beim Lesen der Template-Liste

  /templates/upload:
    post:
      summary: Upload eines ZIP-Templates
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                templateZip:
                  type: string
                  format: binary
                templateName:
                  type: string
      responses:
        200:
          description: Erfolgreicher Upload
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  path:
                    type: string
        400:
          description: Fehlende Datei oder Template-Name
        500:
          description: Fehler beim Entpacken des ZIP-Archivs